[Feat] The CommunityEvents Base , Refactoring The Prev Code
Why: so we could use same component in main page and in the admin Dashboard
How: use shared component
