+++ src/components/Common/GeneralDashboardSidebar/index.js
@@ -1,0 +1,2 @@
+import GeneralDashboardSidebar from "./GeneralDashboardSidebar";
+export default GeneralDashboardSidebar;
