+++ src/components/CommunityEvents/index.js
@@ -1,0 +1,2 @@
+import CommunityEvents from "./CommunityEvents";
+export default CommunityEvents;
