+++ src/components/Opportunities/DisplayCommunityEvents/DisplayCommunityEvents.jsx
@@ -1,0 +1,15 @@
+import React from "react";
+
+import CommunityEvents from "../../CommunityEvents/CommunityEvents";
+
+const DisplayCommunityEvents = () => {
+    return (
+        <CommunityEvents
+            pageHeader
+            title="Community Events"
+            subtitle="Join to any of the Community Events to develop your skill set."
+        />
+    );
+};
+
+export default DisplayCommunityEvents;
