+++ src/components/Opportunities/DisplayCommunityEvents/index.js
@@ -1,0 +1,2 @@
+import DisplayCommunityEvents from "./DisplayCommunityEvents";
+export default DisplayCommunityEvents;
